+++ tests/test_tools.py
@@ -92,6 +92,21 @@ def test_solve():
     cauchy = fem.tools.project(fem.math.tovoigt(s), region=r, mean=True)
     assert cauchy.shape == (r.mesh.npoints, 6)
 
+    cauchy = fem.tools.topoints(s, region=r, sym=False)
+    assert cauchy.shape == (r.mesh.npoints, 9)
+
+    cauchy = fem.tools.topoints(s, region=r, sym=True)
+    assert cauchy.shape == (r.mesh.npoints, 6)
+    
+    cauchy = fem.tools.topoints(s[:2, :2], region=r, sym=True)
+    assert cauchy.shape == (r.mesh.npoints, 3)
+    
+    cauchy = fem.tools.topoints(s[:2, :2], region=r, sym=False)
+    assert cauchy.shape == (r.mesh.npoints, 4)
+
+    cauchy = fem.tools.topoints(s[0, 0], region=r, mode="scalar")
+    assert cauchy.shape == (r.mesh.npoints,)
+
 
 def test_solve_mixed():
     r, f, fields = pre()
@@ -299,6 +314,160 @@ def test_newton_body():
     )
 
 
+def test_project():
+    # rectangle (triangle)
+    mesh = fem.Rectangle(n=2).triangulate()
+    region = fem.RegionTriangle(mesh)
+    field = fem.FieldAxisymmetric(region, dim=2)
+    values = field.extract()
+
+    projected = fem.project(values, region, average=False)
+    assert projected.shape == (mesh.cells.size, 3, 3)
+    assert not np.any(np.isnan(projected))
+    assert np.all([np.allclose(np.eye(3), res) for res in projected])
+
+    projected = fem.project(values, region, average=True)
+    assert projected.shape == (mesh.npoints, 3, 3)
+    assert not np.any(np.isnan(projected))
+    assert np.all([np.allclose(np.eye(3), res) for res in projected])
+
+    projected = fem.project(values, region, average=True, mean=True)
+    assert projected.shape == (mesh.npoints, 3, 3)
+    assert not np.any(np.isnan(projected))
+    assert np.all([np.allclose(np.eye(3), res) for res in projected])
+
+    # rectangle (quadratic triangle)
+    mesh = fem.Rectangle(n=2).triangulate().add_midpoints_edges()
+    region = fem.RegionQuadraticTriangle(mesh)
+    field = fem.FieldAxisymmetric(region, dim=2)
+    values = field.extract()
+
+    projected = fem.project(values, region, average=True, mean=True)
+    assert projected.shape == (mesh.npoints, 3, 3)
+    assert not np.any(np.isnan(projected))
+    assert np.all([np.allclose(np.eye(3), res) for res in projected])
+
+    # this is wrong
+    projected = fem.project(values, region, average=True)
+    assert projected.shape == (mesh.npoints, 3, 3)
+    assert not np.any(np.isnan(projected))
+    with pytest.raises(AssertionError):
+        assert np.all([np.allclose(np.eye(3), res) for res in projected])
+
+    # cube (tetra)
+    mesh = fem.Cube(n=2).triangulate()
+    region = fem.RegionTetra(mesh)
+    field = fem.Field(region, dim=3)
+    values = field.extract()
+
+    projected = fem.project(values, region, average=False)
+    assert projected.shape == (mesh.cells.size, 3, 3)
+    assert not np.any(np.isnan(projected))
+    assert np.all([np.allclose(np.eye(3), res) for res in projected])
+
+    projected = fem.project(values, region, average=True)
+    assert projected.shape == (mesh.npoints, 3, 3)
+    assert not np.any(np.isnan(projected))
+    assert np.all([np.allclose(np.eye(3), res) for res in projected])
+
+    projected = fem.project(values, region, average=True, mean=True)
+    assert projected.shape == (mesh.npoints, 3, 3)
+    assert not np.any(np.isnan(projected))
+    assert np.all([np.allclose(np.eye(3), res) for res in projected])
+
+    # cube (quadratic tetra)
+    mesh = fem.Cube(n=2).triangulate().add_midpoints_edges()
+    region = fem.RegionQuadraticTetra(mesh)
+    field = fem.Field(region, dim=3)
+    values = field.extract()
+
+    projected = fem.project(values, region, average=True, mean=True)
+    assert projected.shape == (mesh.npoints, 3, 3)
+    assert not np.any(np.isnan(projected))
+    assert np.all([np.allclose(np.eye(3), res) for res in projected])
+
+    # this is wrong
+    projected = fem.project(values, region, average=True)
+    assert projected.shape == (mesh.npoints, 3, 3)
+    assert not np.any(np.isnan(projected))
+    with pytest.raises(AssertionError):
+        assert np.all([np.allclose(np.eye(3), res) for res in projected])
+
+
+def test_extrapolate():
+    # rectangle (triangle)
+    mesh = fem.Rectangle(n=2)
+    region = fem.RegionQuad(mesh)
+    field = fem.FieldAxisymmetric(region, dim=2)
+    values = field.extract()
+
+    projected = fem.tools.extrapolate(values, region, average=False)
+    assert projected.shape == (mesh.cells.size, 9)
+    assert not np.any(np.isnan(projected))
+    assert np.all([np.allclose(np.eye(3).ravel(), res) for res in projected])
+
+    projected = fem.tools.extrapolate(values, region, average=True)
+    assert projected.shape == (mesh.npoints, 9)
+    assert not np.any(np.isnan(projected))
+    assert np.all([np.allclose(np.eye(3).ravel(), res) for res in projected])
+
+    projected = fem.tools.extrapolate(values, region, average=True, mean=True)
+    assert projected.shape == (mesh.npoints, 9)
+    assert not np.any(np.isnan(projected))
+    assert np.all([np.allclose(np.eye(3).ravel(), res) for res in projected])
+
+    # rectangle (quadratic triangle)
+    mesh = fem.Rectangle(n=2).add_midpoints_edges()
+    region = fem.RegionQuadraticQuad(mesh)
+    field = fem.FieldAxisymmetric(region, dim=2)
+    values = field.extract()
+
+    projected = fem.tools.extrapolate(values, region, average=True, mean=True)
+    assert projected.shape == (mesh.npoints, 9)
+    assert not np.any(np.isnan(projected))
+    assert np.all([np.allclose(np.eye(3).ravel(), res) for res in projected])
+
+    # this is wrong
+    with pytest.raises(ValueError):
+        projected = fem.tools.extrapolate(values, region, average=True)
+
+    # cube (tetra)
+    mesh = fem.Cube(n=2)
+    region = fem.RegionHexahedron(mesh)
+    field = fem.Field(region, dim=3)
+    values = field.extract()
+
+    projected = fem.tools.extrapolate(values, region, average=False)
+    assert projected.shape == (mesh.cells.size, 9)
+    assert not np.any(np.isnan(projected))
+    assert np.all([np.allclose(np.eye(3).ravel(), res) for res in projected])
+
+    projected = fem.tools.extrapolate(values, region, average=True)
+    assert projected.shape == (mesh.npoints, 9)
+    assert not np.any(np.isnan(projected))
+    assert np.all([np.allclose(np.eye(3).ravel(), res) for res in projected])
+
+    projected = fem.tools.extrapolate(values, region, average=True, mean=True)
+    assert projected.shape == (mesh.npoints, 9)
+    assert not np.any(np.isnan(projected))
+    assert np.all([np.allclose(np.eye(3).ravel(), res) for res in projected])
+
+    # cube (quadratic tetra)
+    mesh = fem.Cube(n=2).add_midpoints_edges()
+    region = fem.RegionQuadraticHexahedron(mesh)
+    field = fem.Field(region, dim=3)
+    values = field.extract()
+
+    projected = fem.tools.extrapolate(values, region, average=True, mean=True)
+    assert projected.shape == (mesh.npoints, 9)
+    assert not np.any(np.isnan(projected))
+    assert np.all([np.allclose(np.eye(3).ravel(), res) for res in projected])
+
+    # this is wrong
+    with pytest.raises(ValueError):
+        projected = fem.tools.extrapolate(values, region, average=True)
+
+
 if __name__ == "__main__":
     test_solve()
     test_solve_mixed()
@@ -308,3 +477,5 @@ if __name__ == "__main__":
     test_newton_plane()
     test_newton_linearelastic()
     test_newton_body()
+    test_project()
+    test_extrapolate()